Refactor CurveValuesView.tsx
