+++ Src/WitsmlExplorer.Frontend/components/ContentViews/table/tableParts.ts
@@ -1,4 +1,6 @@
 import React from "react";
+import { WITSML_INDEX_TYPE_DATE_TIME } from "../../Constants";
+import { CurveSpecification } from "../../../models/logData";
 
 export interface ExportableContentTableColumn<T> extends ContentTableColumn {
   columnOf: T;
@@ -82,4 +84,41 @@ export const getCheckedRows = (currentRow: ContentTableRow, data: ContentTableRo
   const shouldCheck = checkedContentItems.findIndex((checkedItem) => checkedItem.id == currentRow.id) === -1;
   const checkedRows = updateCheckedRows(checkedContentItems, rowsInRange, shouldCheck);
   return checkedRows;
+};
+
+export const getProgressRange = (startIndex: string, endIndex: string, indexType: string) => {
+  return indexType === WITSML_INDEX_TYPE_DATE_TIME
+    ? {
+        minIndex: new Date(startIndex).getTime(),
+        maxIndex: new Date(endIndex).getTime()
+      }
+    : {
+        minIndex: Number(startIndex),
+        maxIndex: Number(endIndex)
+      };
+};
+
+export const updateProgress = (index: string, minIndex: number, maxIndex: number, indexType: string) => {
+  const normalize = (value: number) => ((value - minIndex) * 100) / (maxIndex - minIndex);
+  if (indexType === WITSML_INDEX_TYPE_DATE_TIME) {
+    return normalize(new Date(index).getTime());
+  } else {
+    return normalize(Number(index));
+  }
+};
+
+export const getColumnType = (curveSpecification: CurveSpecification) => {
+  const isTimeMnemonic = (mnemonic: string) => ["time", "datetime", "date time"].indexOf(mnemonic.toLowerCase()) >= 0;
+  if (isTimeMnemonic(curveSpecification.mnemonic)) {
+    return ContentType.DateTime;
+  }
+  switch (curveSpecification.unit.toLowerCase()) {
+    case "time":
+    case "datetime":
+      return ContentType.DateTime;
+    case "unitless":
+      return ContentType.String;
+    default:
+      return ContentType.Number;
+  }
 };
